Update MainCode
I took off one library (servo.h) it is because pin 10 and 9 will corrupt and we cant use analogWrite on the program. It will be converted into digitalWrite. What i did for today is to make a new method that replaces the servo.h for the neck of the bot. All of the pins are now working.